backend/requirements.txt
fastapi
uvicorn
python-socketio
python-dotenv
docker
requests
jinja2
sqlalchemy
python-jose[cryptography]
bcrypt
python-multipart
